backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
SQLAlchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
PyJWT==2.9.0
passlib>=1.7.4,<2
cryptography
